+++ crates/stdio_server/src/session/providers/filer.rs
@@ -77,8 +77,6 @@ pub fn read_dir_entries<P: AsRef<Path>>(
     Ok(entries)
 }
 
-pub struct FilerSession;
-
 #[derive(Clone)]
 pub struct FilerMessageHandler;
 
@@ -103,10 +101,12 @@ impl HandleMessage for FilerMessageHandler {
                 }
             }
             // TODO: handle on_typed
-            RpcMessage::OnTyped(msg) => handle_message(msg),
+            RpcMessage::OnTyped(msg) => handle_custom_message(msg),
         }
     }
 }
+
+pub struct FilerSession;
 
 impl NewSession for FilerSession {
     fn spawn(&self, msg: Message) -> Result<Sender<SessionEvent>> {
@@ -120,7 +120,7 @@ impl NewSession for FilerSession {
         };
 
         // handle on_init
-        handle_message(msg);
+        handle_custom_message(msg);
 
         session.start_event_loop()?;
 
@@ -128,7 +128,7 @@ impl NewSession for FilerSession {
     }
 }
 
-pub fn handle_message(msg: Message) {
+pub fn handle_custom_message(msg: Message) {
     let cwd = msg.get_cwd();
     debug!("Recv filer params: cwd:{}", cwd,);
